chore: Export ClerkWebAuthnError from shared error package
